recover from panics from procs (#473)
If a panic occurs during proc runtime, recover from it inside the mux, and report it as an error. I verified by adding support to ztest for an expected error regex, and trigger an integer division by zero panic with put.

* add panic recovery to mux run

* support error regex checking for ztest

* add test for integer divide by zero in put

* make ztest run return separate output and warnings-or-error strings

When ztest run() operates using the zq binary, any warnings generated
will get mixed in with an error message due to the zq command failing.

I expect tests will fall into 3 categories, depending on what they
want to verify: a) just the output, b) output and warnings,
or c) an error message.

This change adds a return value to run() that holds warnings when no
error is reported, or the error string when there is one. This makes it
possible to write tests that fall into one of those 3 categories, _and_
works regardless of whether run() uses the zq binary or evals the zql
programatically with the mux.