Tweaks to put warning handling
The put example in the zql docs was causing a warning since some of the
logs in the corelight sample have "orig_bytes" unset.  Change the example
in the docs to use -q to ignore the warning (since it is a distraction
in the docs) but also improve the warning in this case.