Graceful http shutdown
Add pkg httpd. httpd.Server is an http Server that provides user-friendly
graceful http shutdown. The Server instance's Run and Start method both take
a context that will be used as the base context for all requests issued
initiated by the server. If the passed context is cancelled,  the Server
closes the active listener and waits for all active connections to become
idle. Because the passed context is used base context for all requests,
when the context is cancelled all active requests in turn receive a
cancellation signal. If the wait for active connections to idle exceeds
the duration set in httpd.ShutdownTimeout the Server forcefully closes all
active connections.

It is up to the implmentor of each handler route to determine whether to
listen to a request's ctx.Done() signal. It is in this pr author's opinion
that short-lived requests (e.g. space info, space list, status, etc) should
ignore ctx.Done() and all long running requests (e.g. searches and data posts
to spaces) should start to exit on ctx.Done(). This will the effect that during
graceful shutdown short-lived requests will be allowed to finish to completion,
while long running requests will terminate and prevent a forced shutdown.

Additional changes:
- A pkg catcher: a user friendly interface for catching signals in
a user friend manner.
- Have zqd listen utilize pkg catcher and httpd to trigger a graceful
shutdown when an interrupt signal is received.