make space deletion cancel and await other operations (#451)
When a space delete API call is made, signal other active operations to
cancel and wait for them to do so, preventing further operations
until the deletion is done.

This uses zqd.Core as the holder of the necessary state. Eventually, I
think this would better live in a dedicated space manager, but that's a
much larger change, and I think this code will move there easily.

This is a part of addressing ingest cancellation during application
shutdown on Windows, as we currently see "file in use" errors there when
a delete is called concurrently with ingest.