Crude prototype for zng archive (zar) (#482)
This commit adds a zar command for creating and searching
bzng index files.  The index files are organized as sorted
string tables and use the sst package.

This commit is a work-in-progress and does not impact zq or zqd.

This zar prototype is very early and and currently only
indexes IP addresses as a hard-wired configuration.  The code
however is set up to do more flexible indexing soon.

The sst package uses a simple binary format for framing
keys and values as arbitrary byte sequences but we realized
it could be a very powerful foundation to store the sst
files as bzng.  We will implement this change in a future PR.