Driver refactor (#549)
This change introduces a new driver.Driver interface consisting of methods needed to handle the output of running a flowgraph. It is derived from the existing driver.Driver struct, which is now driver.CLI. In addition, there is now a single func to run a flowgraph (driver.Run(..)), instead of the two slightly different ones we had before.

All places that execute flowgraphs are ported to use a driver.Driver and execute the flowgraph via driver.Run(). This includes search, ingest, zq, and tests.