Fix error handling in search endpoint handler
Previously, the search endpoint could call http.Error() after
having sent data. This commit fixes that by breaking the search action
into a setup phase and a run phase.

If the setup phase fails, an HTTP error is returned. If the setup
phase succeeds, a 200 is returned, and  any error encountered during
the execution of the search will result in a TaskEnd message with the
error being sent to the client.