add zbuf.Peeker
This commit adds a generic zbuf.Peeker that wraps a zbuf.Reader and allows you
to inspect the next zng.Record in the stream without actually reading it.

This code came from the port of sst to zng. It is not currently used anywhere in the
zq repo but is a useful, general functionality for the framework.