zq: new "continue with input errors" flag
This PR adds the `-e` flag to zq, which controls the behavior of zq
when reading from multiple files and encountering a read or parser
error on one of those files.

The default behavior is unchanged: an input error on any input file
causes processing to stop on all inputs.

If -e=false, upon encountering an error, zq emits a warning, closes
the error'd file, and continues reading from any other readers.

To support this, some plumbing (which will be used by zqd soon) had to
be done along the way:

- Make scanner.NewScanner take a *proc.Context (like all other procs) and
configure the proc context's warning channel on the scanner's reader
(if an appropriate config method is present).

- Add said config method to scanner.Combiner, and change Combiner's read
loop to emit the warning upon error in one of its readers, if a
warnings channel has been configured.