JSON: error on incomplete descriptors
Previously, incomplete descriptors did not lead to any user-visible
notification. (A stats counter is increased, but flowgraph stats are
not currently hooked up to anything, see #555).

This PR turns incomplete descriptors into errors (like other json type
errors), which will lead to their notification in the Brim app and on
the CLI.