add Localize method to resolver.Context
This commit adds a method too resolver.Context to localize
a record type from another type context.  This emphasize
the design pattern here where all types live within a
type context and if you have a foreign type it is easily
tranlated into the local context, and it's important to do so.