Re-work Search (#597)
Reintroduce an AST node for search terms. We used to have a SearchString AST node but it was removed in #174. Since then, the AST for searches has slowly grown more unwieldy. This had several disadvantages:

    it was encoding the implementation of search rather than representing the syntax of a query
    it required clumsy new additions to other parts of the AST (e.g., {op: CompareField, Comparator = "searchin"})
    it was inefficient since some searches ended up OR-ing together 3 or 4 clauses which ended up making 3 or 4 passes over all the fields of each record

Address this all by creating a simple Search AST node and moving the logic for search into the filter compiler.