Introduce =~ operator in filters
This new operator means roughly "does the thing on the left hand side
match the pattern on the right hand side".  It is used for glob and
regular expression matching and for matching addresses against subnets.
So, the filter "foo = abc*" now matches the literal string "abc*", glob
matching is expressed as "foo =~ abc*".  A "naked" search for a glob
expression still works as before.
Filters that previously used the = operator for matching are no longer
valid, various tests and docs examples were updated to reflect the new
requirement.