refactor file reading/writing so other tools can leverage (#608)
This commit splits zio.Flags into ReaderFlags and WriterFlags and refactors the file opening code so it lives in one place. This will make it easier to re-use file logic with command options by other tools in the zq package.

The handling of stdin in detector.OpenFile is goofy and this will be fixed in a subsequent PR.

Another next step is to move much of the json typings stuff into zio and come up with some new names to reflect the fact that this is solving the schema mapping problem.