extend ztest to handle arbitrary shell scripts (#607)
This commit extends the yaml-based ztest framework to handle
arbitrary scripts.  The details are documented in the ztest package
comment.  The input/output specs have been generalized a bit
and we may want to refactor this all a bit more so that the
newer bits of funcionality can be used by the existing zq-direct
commands and in-process runs.

Only a couple of the old go tests under tests/suite have been
ported over in this commit.  Once this commit is reviewed and
merged, we can port the rest over in a subsequent PR.