open output files with truncation (#632)
This commit fixes a bug where the zng.Writer output files
were open without truncation.  This would cause corrupt
files to be created when overlaying new but shorter output
on an existing file.  This happened in a few places.
Also, we changed O_RDWR to O_WRONLY in a couple places 
that files were used write only.