add -x to zdx create to handle already-sorted key,val zng tables
This commit adds a flag to "zdx create" to take as input a zng stream
that is already formatted into sorted key records but is not in the
form of an zdx index file with related b-tree sections.

This is useful for running zql queries to create arbitrarily rich
indexes that can then be wrapped into a zdx index with "zdx create -x".