fix crash when index file is present but empty (#650)
This commit fixes a bug where the keyType returned by finder.Open()
is nil. This happens when an index exists but has nothing in it.
The fix is to return a false match in this case.  Also, added a test 
to cover this.

We also noticed the finder wasn't being closed here leaving
potentially many files opened during a traversal.  We added
logic to close the temporary finder used in each search.