add "zar zq"
This command adds support for the "zq" sub-command of zar, which
lets you run zq-style commands over the log files in a zar archive.
In this case, the file "_" is the log file in question and any
files referred to live within the zar directory associated with
the log.  The zq command is run for each log file found.