allow baselayer zdx index to have arbitrary values (#684)
This commit allows the base-layer zdx index to have arbitray
structure, requiring only that there is a field called "key"
and that the records are sorted by the key.  When we change
the zdx file format to be embodied in a single zng file, we
will relax this requirement, allowing the key to be any field,
and will put the key field's name in the header of the zng
index file.

This change allows us to put interesting information inside indexes,
e.g., analytics aggregations, so that computation can be carried
across index files by piping the output of a zar search into
a zq command that can aggregate or otherwise process the
information stored in the row-slices of the index.