allow zar find to stream hits as zng
This commit changes zar find so if you specify an output file
or ("-o -" for stdout) zar find will output zng instead of a list
of log paths, where hte zng file is comprised of a stream zng.Records
where each record corresponds to the records with the matched key
for the hit from each index file.

For now, this is kind of hacked together for prototyping and demo
purposes.  We will clean it up soon a forthcoming PR.