support multi-key indexes and clean up zar/zdx
This commit adds supprt for multi-key indexes and refactors
significant pieces of the zar/zdx machinery.

The zar index files are still stored as separate files but
a header has been added to the base layer to provide magic,
version, key fields, and customization of the btree child
field (since it is stored alongsiide the key fields in the
index and there can be a name collision).

We also reworked index searches to return result over a
channel so that results from a dir walk are easily managed.

A bug is fixed where null values would end up in the index.
It does not make sense to index non-existant values.

We separated the zdx create command into create and convert,
where the latter converts an existing sorted-key zng file
into a zdx and the former creates a sorted-key table using
an in-memory map (this should be replaced eventually with
zql proc code when we have a spill-capable group-by).

There is a new "zdx seek" command, which indexes just
first records after the start-of-stream (sos), where the
value alongside each key is the seek offset of the sos.

We also made the behavior of zar find more sensible with the
-z option, when selecting between file-path output and zng output.

To make multi-key search work properly, we added a new comparator
function, expr.KeyCompareFn, which stores the key in the
closure and returns a comparison result for an arbitrary record
with the stored key.  In this case, the key and the record
can be of different types and the respective key fields in
no consitent order between the types.