Have scanner listen and stop on context.Cancel (#709)
The lack of context listening to the cancel signal was causing a
a long running ingest handler to not exit when the context of the
request was cancelled. Add context.Err check to scanner which lets
the flowgraph elegantly cease operations.

Also there was a deadlock occurring in the split proc when the parent
proc communicates EOS and the split proc has receieved no data.
The was occurring because the child SplitChannel Procs where themselves
listening for context.Done() and exiting before the SplitProc could
tell them itself. Tell the SplitChannel to calm down and trust the
process.

closes #707