fix zng reader to properly handle stream reset
This commit fixes a bug in the zng reader where we were not properly
tracking the internal type context separately froom the passed-in shared context.
The fix is to keep an internal context along with the external context
and teardown the mapper that maos between them whenever the context
is reset.

We also added a test that exercises this case by encoding two different
values with a max-stream size of 1 record so the two records will be
different types but have the same type ID in the mapper table.