add zng.Builder.Parse()
This commit adds a Parse method to the zng builder so that records
can be built from a DFS traversal of the string representation of
the leaf values.  A future PR will define a syntax for record values
and do a proper recursive descent here, but for now, this will be
used by the upcoming zar/zdx PR for multi-value keys.