spawn zeek on windows with new hidden subcommand (#717)
This adds a hidden zqd subcommand "winexec", that executes a command
using the same job object based spawned process termination we are
already using in the windows zeek-launcher utility. Use this subcommand
when we launch zeek on windows. This is a first step to treating zeek as
more of a black box from inside zqd, so that users can run their own
zeek setup.