Add space PUT endpoint
For endpoint only allows users to update the name of the space. As
with other space accessors this method is not thread safe. This
will be fixed soon, but I wanted to get the feature in to allow
brim devs to fully implement space name change functionality in the
interim.