add zng.Builder.Parse() (#730)
This commit adds a Parse method to the zng builder so that records
can be built from a DFS traversal of the string representation of
the leaf values.  A future PR will define a zql syntax for all types of 
zng literals and either do a proper recursive descent here or use 
the peg parser to generate an AST that can be translated into zng.Records.