Space Storage (#761)
Make package Storage, responsible for managing all log data for a space.

Eventually there should be a central interface for Storage so that
a space can support multiple storage backends. For now this is limited
to a single ZngFile storage backend where:
- Data is populated by passing a zbuf.Reader into it.
- Data can only be rewritten. It cannot be updated.
- Data is housed in a single all.zng file.
- Concurrent reads are allowed but only one active write is allowed.
- Maintains an active cache of the total time span of all records stored.

Additional supporting changes:
- Reworked ingest.Log to instead return a LogIngestOp struct that exports
go channels for "Warnings" and "Done" notifications. Also includes
a Stats() function that provides a thread-safe snapshot of current ingest
stats.
- Reworked ingest.Pcap to use space.Storage for generating log snapshots.