Space IDs (#721)
Update space api to support immutable space IDs. The space uses
segmentio/ksuid to generate a date sortable uuid. The space id is
also prefixed by the first 6 alphanumeric characters of the name
or post data path base name.

The space.ID is the name of the space's directory in the zqd root.

There is now no constraints on the characters that can be in
space.Name. Additionally space.Name does not have to be unique
among other spaces.

In order to support backwards compatibility the names of all
previously existing spaces are now the ids of the space.

Also:
- Change spaces list api to return a list of api.SpaceDesc structs
instead of just a list of space names.
- Add space put endpoint. For endpoint only allows users to update
the name of the space. As with other space accessors this method is
not thread safe. This will be fixed soon, but I wanted to get the
feature in to allow brim devs to fully implement space name change
functionality in the interim.

closes #687 #686