Serialize zero spans as a json null
A span with Dur == 0 is an impossible time span- why not use this
as the nil value of span?

Change json serialization of `nano.Span`s so that if span.Dur == 0
the span gets a null value in json. Like was json null values set a
span in a struct to nano.Span{0, 0}.