zapi: Signal based context
For the zapi add a context to the root command that cancels when
SIGTERM or SIGINT is recieved. This context should be passed into
any long running tasks launched by the command. This functionality
will come more into play when the longer running search or post data
commands are added in forth comming prs.