Fix logic for detecting the end of lists
The logic for determining the end of a list based on the RL field was
completely wrong and just working by accident.  Fix it here.
Ideally this should have a test, a good test for this would read a
parquet file that contains a LIST-valued column where all rows in the
file have exactly 1 element in the list.