Add options to ParquetReader (#830)
The column handling can now be customized -- columns with unknown structure
that the reader cannot handle can either be a fatal error (the default) or
ignored.  Also, a specific list of column names can be passed in in which
case the reader will not do the I/O and decoding work for other columns.
None of these options are yet wired up to zq command line arguments.