Let expr tests be written with nil records (#835)
Every expression is evaluated in the context of a zng record, but a bunch
of our tests just evaluate literal expressions so requiring that they
create a record makes the tests more verbose and harder to follow.  Fix
the test helpers to allow for a nil record and clean up affected tests.