detector.MultiFileReader
MultiReader returns a zbuf.ReadCloser that's the logical concatenation of the
provided input paths. They're read sequentially. Once a file has reached EOF,
Close is called on the file. Once all inputs have reached end of stream,
Read will return end of stream. If any of the readers return a non-nil error,
Read will return that error.

The primary use case is for reading large amounts of files but
limiting the amount of open file descriptors.

Use MultiFileReader when doing zar searches in zqd as well as
performing searches via the zar command.

closes #832