zq command: allow output to s3 bucket
Add package s3io with Writer struct- a wrapper around s3manager
that implements the io.WriteCloser interface. This allows an s3 bucket
to be used interchangeably with other output sources such as a file
or stdout.

Check path in emitter.NewFile, if the path is a valid s3 path s3io.Writer
is used.

On the testing front I was most curious about error handling for
s3io.Writer. I looked at using minio as is done for s3 capabilities
in detector but found this approach not conducive to creating tests
with error cases. As a result I utilized golang/mock to generate
a mock s3.client and construct simple tests around a couple key error
cases.

Also:
- gomock relies on go generate to regenerate mock structs. As a result
add make generate command which runs generate on code base and returns
exit code if a git diff is found.

closes #771