Decomposable reducers (#892)
Add a ConsumePart() and ResultPart() methods and implement them in reducers that perform "decomposable aggregations".

And a couple supporting changes:

* Remove record requirement for reducer instantiation

Originally, reducer instantiation did not require a record. Then we
added that requirement in order to support some corner cases with
missing values. But those corner cases can be handled without the
initial record, and this change does that.

* Remove reducer.Result()

Call the reducer's Result() method directly instead.