Add package iosource
Add package iosource which defines a registry for sources- A source
is way of opening reading/writers for paths with different schemas.
For the zq command add s3 source to the default registry.

Use a mock source for an iosource.Registry in order to test s3
path support for emitter.Dir.