s3io: strip starting slash from key (#1032)
This was causing an issue on the real s3 api for the Prefix property
of the delete batch and list batch api methods. Works for minio's
mock s3 implementation, not for the real s3 api. As it turns out
the starting "/" is not needed for keys, just strip it in
s3io.parsePath.

Closes #1026 & #1024