zapi pcappost: fix -f flag when space already exists (#1042)
If a pcappost command is run on with the -f (force) flag on a space
that already exists we get a nil pointer panic. Fix this.

Also:
- Add zeek as a dependency of system tests and run zqd system tests
that depends on zqd runner.
- In the zqd listen command if the -zeekrunner arg is not set, check
if zeekrunner command can be found in $PATH. In order to get zeek
system tests running this had to be done.