impose limit on the number of levels in a microindex (#1119)
This commit adds a limit on the number of levels that comprise the
index-key tree in a microindex.  Since each node is relatively large
(e.g., 1000s of keys), this limit will only be hit if the frame threshold
is made small enough to limit the number of keys per level to 1.
Even with 2 keys per level, we can represent 2^MaxLevels keys in the
base layer, e.g., a million or so keys, but more typically with
say 512 keys per level, we'd be at 512^20 = 2^200.