Configurable merge (#1121)
* Configurable merge

Add flags to ParallelProc to indicate ordering behavior when merging
the branches. The MergeOrderField contains the name of the field in
which outputs of each parallel branch are ordered. If it is
zero-valued, then the existing (unordered) merge proc is used.

These flags not currently exposed in ZQL syntax but could be at a
later point.

* feedback