add warnings channel to pcap-ng reader and report through zqd api
This commit works around a bug in certain pcap-ng files where extra
bytes show up at the end of the file that are not a compliant pcap-ng
block.  The fix is to log a warning rather than throw an error.

To add warnings, we hacked in some hooks to the pcap-post endpoint.
This should be cleaned up and normalized with log-post endpoint
in a later commit.