add support for empty microindexes (#1136)
This commit adds support for the notion of a valid microindex with
nothing in it.  This is useful for knowing when something has been
indexed but there are no matching elements.  An empty microindex
is a zng file with a microinidex trailer with various null fields.

We also fixed a bug where zar index would leave behind an empty file
(which is not a valid microindex) when it encountered an error.
The fix is to add an Abort() method on the microindex Writer that can
be called in place of Close() to cleanup any partial state.