create zst format
This commit introduces the zst columnar format to zq.  Package zst
provides the core new functionality.  Package zstio is a zio wrapper
for zst to wire it into the zq pipeline.  A new CLI command "zst"
provides a tool for test, debug, and demo.

There is much more work to be done for columnar format but this
should serve as a good foundation to start from.