fix bug in put proc and refactor (#1158)
The put proc had a bug where it attempted to append a clause result
after all of the clauses had already replaced fields in the original
input.  Put.put() was doing many different things with some confusing
variable and type names so a refactor was warranted.

This commit fixes the bug and refactors the logic into bit-sized chunks
with hopefully more helpful names.  The resulting code is smaller.