driver.parallelizeFlowgraph: put head/tail before and after merge (#1182)
x | tail was previously parallelized into (x; x; ...) | tail, which
meant that all records had to be pulled through the ordered merge. It
can instead be parallelized into (x | tail; x | tail; ...) | tail so
that only the tails go through the ordered merge. The same is done for
head for consistency, though the performance implications there are
probably not significant.