cut -c: use descriptor to derive fields (#1227)
* cut -c: use descriptor to derive fields

Previously, cut -c used a zng.fieldIter on the first record of a type
to compute which fields to extract. This leads to problems with unset
values: if an inner record is unset, fieldIter doesn't visit leaves,
and the cutter gets confused, as observed in https://github.com/brimsec/zq/issues/1226.

Instead, use the record's descriptor to figure out which fields to cut.