change union implementation to conform to zng spec (#1245)
Union values were being incorrectly encoded by double encoding the
length of the selector integer.  This commit changes things to
conform.

We also noticed a bug in the ndjson reader where the union array
values were being encoded as a sequence selector, value, selector,
value instead of sequence of containers of selector/value pairs, 
so we fixed this too.