workround pcap filtering problem with zeek's truncated timestamps (#1243)
Zeek truncates packet timestamps to microseconds so packets
that have nanosecond-precison timestamps can fall outside of
the range of the the implied microsecond-precision time span
of a conn record.  This commit puts a hack in the pcap search
endpoint to round adjust the requested span to account for
the truncation of both the conn timestamp and its duration.