fix bug where bad record type allowed through (#1260)
This commit fixes a bug where a field name discovered by
searching for a colon could have illegal characters in it,
which should result in a syntax error but happened to be
interpreted as a valid type.

The fix is to check that field names discovered by the colon
search conform with the type syntax.  A better fix would be
to use peg or write a proper recursive descent parse for
the zng type system  Maybe later.

While we were at it, we noticed there was a problem where extra
text after the end of a valid type would be ignored instead of
flagging an error.