update zjson spec and implementation
This commit updates the zjson spec to conform with feedback from
our experience and use.  The main change is to format types
completely using structured JSON instead of a mix of structure
for record types and strings for array/set/union.  Likewise,
we fixed the aliases to be coded as structurerd JSON instead
of type strings.

Also, joe needed an Any type to be able to code something in a struct
that could either be a string or an object.