update zapi to package zq/cli and other minor changes (#1278)
This commit updates "zapi get" to use the output flags options
in package zq/cli.  It also changes the "protocol" option to
"encoding" representing the server format encoding of the http
response.  This sets up "zapi get" to handle the vanilla ndjson
and csv support that is come to the zqd search end point.

We also updated the online help a bit.