add dataframe proc
This commit adds a dataframe proc "df" which transforms the input
zng stream into an output stream that all has the same uber schema
by taking the union of all of the record types in the input and
coding each record to its position in the uber schema.

Because df needs to know all the types to build the uber schema,
it must buffer all of its output before it can generate the first
output record, just like sort  Thus, there is a memory-limit
circuit breaker.

The next step here is to add logic to spill to disk when the breaker
is tripped.  This is easier than sort because you don't need multiple
runs in different files and a merge.  It can all be done with a
single file.