Suricata launcher (#1294)
This commit adds support for running Suricata as part of the pcap ingest process and importing the resulting eve.json alongside the zeek logs. This is just the first step of the overall feature. Notably still missing are (a) packaged suricata binaries that can be installed via the makefile like Zeek and (b) proper conversion of eve.json including types and renaming the timestamp field.