one more tmie fixing CLI flags refactor
The last refactor broke "zq -version".  The previous logic was a bit
fragile because cli.Flags.Init() would return a flag to stop
executing if -version was given.  But now, Init() can fail for
other reasons, e.g., writing binary data to stdout.  This would
block the help screen if Init was before the help check, but if we
put it after, then help would be displayed with -version.

The fix is to simplify the logic by cli.Init calling os.Exit if
"-version" is given and putting the error check after the subsequent
check for help.  This only applies to zq and not the other tools
because the other tools use subcommands and the issue does not
arise with the root command that does not have outputFlags checks.

We also added procFlags to zqd since they were missing in the last
refactor.  This gives "-sortmem" back to zqd.