fix cleanup of groupby spilled files on sigint
This commit fixes a bug where the groupby proc would send end-of-stream
in parallel with cleaning up the spill directory.  This causes a race
between cleaning up and the process exiting and was reproducible by
running a large group-by with zq and ctrl-c'ing the job.

The fix is to finish cleaning up before sending end-of-stream in
response to a Pull.  This will prevent the mux from counting the
upstream proc as done and otherwise exiting too soon.