pcap: Add pcap info command (#1354)
zq#1341 fixed a bug with how pcapio was parsing interface descriptors
for pcapng files. This follow up as an info command along with tests
to ensure the pcap command is correctly gathering pcap info.

This is a simple first pass, not an attempt to exhaustively parse and
display all information available in the pcap|ng file formats.