allow ztest yaml tests to live anywhere (#1160)
This commit lifts the system test driver out of zq/tests/ and into zq/zq_test.go so that ztests may be put anywhere in the source tree.  A ztest is any yaml file that is a descendent of a directory "ztests".  All of the system tests have been moved from tests/{suite,formats,fixedbugs} to zq/ztests so they are captured with the ztests rule and, as an example, the fuse proc tests have been moved to proc/fuse/ztests/*.yaml.